requirements.txt
beautifulsoup4==4.12.2
Django==3.2.16
django-cors-headers==3.12.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-reset-migrations==0.4.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
selenium==4.10.0